* structure refactoring * updated version
