+++ jsr.json
@@ -1,6 +1,6 @@
 {
   "name": "@flatline/js-harvester",
-  "version": "0.3.6",
+  "version": "0.3.7",
   "license": "MIT",
   "exports": "./index.js"
 }
